Add a moby build --pull option to force pulling images
